refactor(rust): split chaoxing class domain module
